requirements.txt
imageio
imageio-ffmpeg
scikit-image
opencv-python
configargparse
tensorboard
tensorboardX
Pillow>=7.2.0
pyyaml>=5.3.1
matplotlib
scipy>=1.6.
omegaconf
tqdm
einops
torch-ema
huggingface_hub
accelerate
diffusers
transformers
peft
# deepspeed
xformers